requirements.txt
pyrogram~=1.2.9
Telethon~=1.23.0
TgCrypto
SQLAlchemy~=1.4.21
psycopg2-binary
pyromod~=1.5
pytz
python-telegram-bot
gunicorn
requests
aiohttp
aiohttp-jinja2
bs4